Extract shared NodeTimeAttribution service to Core
The per-operator "own time" CPU/elapsed calculation was implemented
separately in the desktop plan viewer (PlanViewerControl) and the web
plan viewer (Index.razor), with the desktop version handling exchange
operators via worker-thread data and the web version lacking that.

Consolidate both onto a single Core service (NodeTimeAttribution) ported
from the more-correct desktop implementation, so desktop and web report
identical operator timings. The web's node coloring/timing display now
matches the desktop, including parallelism exchange handling.

The analysis pipeline (PlanAnalyzer.Timing) and the OperatorResult-based
text exporter (TextFormatter) keep their own implementations by design.

NodeTimeAttribution.cs is linked into PlanViewer.Web (Blazor WASM links
Core sources individually rather than referencing the assembly).
